Handle form history overrides in demo userland
