--- packages/ui/.changes/patch.form-navigation-jsx-attributes.md
@@ -1,1 +1,0 @@
-Add a JSX prop type for `data-rmx-history` on submit buttons and inputs.
